Rename `nn` transformer module to `projection`.
